// src/app/dui/models/control-action.ts
export class ActionParams{​​​​​​​​
  actionLink: string;
  closeFormOnFailure: boolean;
  closeFormOnSuccess: boolean;
  messageOnFailure: string;
  messageOnSuccess: string;
  navigateRouteOnFailure: string;
  navigateRouteOnSuccess: string;
  navigateUrlOnFailure: string;
  navigateUrlOnSuccess: string;
  refreshGridOnFailure: boolean;
  refreshGridOnSuccess: boolean;
}​​​​​​​​

export class ControlActions {
  Actions: ActionParams;
  ActionLink: string;
}